feat: create page component
